azure_functions/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
tweepy==4.8.0
azure-data-tables==12.4.0